+++ sqlg-test/src/main/java/org/umlg/sqlg/test/topology/TestTopologyPropertyColumnRename.java
@@ -3,23 +3,17 @@ package org.umlg.sqlg.test.topology;
 import com.google.common.base.Preconditions;
 import org.apache.commons.collections4.set.ListOrderedSet;
 import org.apache.commons.lang3.tuple.Triple;
+import org.apache.tinkerpop.gremlin.process.traversal.P;
+import org.apache.tinkerpop.gremlin.structure.T;
 import org.apache.tinkerpop.gremlin.structure.Vertex;
 import org.junit.Assert;
-import org.junit.Assume;
 import org.junit.Before;
 import org.junit.Test;
-import org.umlg.sqlg.structure.PropertyType;
-import org.umlg.sqlg.structure.TopologyChangeAction;
-import org.umlg.sqlg.structure.TopologyInf;
-import org.umlg.sqlg.structure.topology.PropertyColumn;
-import org.umlg.sqlg.structure.topology.Topology;
-import org.umlg.sqlg.structure.topology.VertexLabel;
+import org.umlg.sqlg.structure.*;
+import org.umlg.sqlg.structure.topology.*;
 import org.umlg.sqlg.test.BaseTest;
 
-import java.util.ArrayList;
-import java.util.LinkedHashMap;
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
 
 public class TestTopologyPropertyColumnRename extends BaseTest {
 
@@ -35,10 +29,9 @@ public class TestTopologyPropertyColumnRename extends BaseTest {
     public void testPropertyRename() {
         TestTopologyChangeListener.TopologyListenerTest topologyListenerTest = new TestTopologyChangeListener.TopologyListenerTest(topologyListenerTriple);
         this.sqlgGraph.getTopology().registerListener(topologyListenerTest);
-        Assume.assumeTrue(!isMariaDb() && !isMysql());
         this.sqlgGraph.getTopology().getPublicSchema()
                 .ensureVertexLabelExist("A", new LinkedHashMap<>() {{
-                    put("column1", PropertyType.STRING);
+                    put("column1", PropertyType.varChar(10));
                 }});
         this.sqlgGraph.tx().commit();
         Optional<VertexLabel> aVertexLabelOptional = this.sqlgGraph.getTopology().getPublicSchema().getVertexLabel("A");
@@ -69,7 +62,7 @@ public class TestTopologyPropertyColumnRename extends BaseTest {
         Assert.assertEquals("column3", propertyNames.get(0));
         this.sqlgGraph.tx().rollback();
 
-        if (isPostgres() || isMsSqlServer()) {
+        if (this.sqlgGraph.getSqlDialect().supportsTransactionalSchema()) {
             propertyNames = this.sqlgGraph.topology().V().hasLabel(Topology.SQLG_SCHEMA + "." + Topology.SQLG_SCHEMA_PROPERTY).<String>values(Topology.SQLG_SCHEMA_PROPERTY_NAME).toList();
             Assert.assertEquals(1, propertyNames.size());
             Assert.assertEquals("column2", propertyNames.get(0));
@@ -78,11 +71,10 @@ public class TestTopologyPropertyColumnRename extends BaseTest {
 
     @Test
     public void testRenameIdentifier() {
-        Assume.assumeTrue(!isMariaDb() && !isMysql());
         this.sqlgGraph.getTopology().getPublicSchema()
                 .ensureVertexLabelExist("A", new LinkedHashMap<>() {{
-                    put("column1", PropertyType.STRING);
-                    put("column2", PropertyType.STRING);
+                    put("column1", PropertyType.varChar(10));
+                    put("column2", PropertyType.varChar(10));
                 }}, ListOrderedSet.listOrderedSet(List.of("column1")));
         this.sqlgGraph.tx().commit();
         Optional<VertexLabel> aVertexLabelOptional = this.sqlgGraph.getTopology().getPublicSchema().getVertexLabel("A");
@@ -116,7 +108,7 @@ public class TestTopologyPropertyColumnRename extends BaseTest {
         column1.rename("_column1PK");
         this.sqlgGraph.tx().rollback();
 
-        if (isPostgres() || isMsSqlServer()) {
+        if (this.sqlgGraph.getSqlDialect().supportsTransactionalSchema()) {
             aVertexLabelOptional = this.sqlgGraph.getTopology().getPublicSchema().getVertexLabel("A");
             Preconditions.checkState(aVertexLabelOptional.isPresent());
             aVertexLabel = aVertexLabelOptional.get();
@@ -125,6 +117,91 @@ public class TestTopologyPropertyColumnRename extends BaseTest {
             identifiers = aVertexLabel.getIdentifiers();
             Assert.assertEquals(1, identifiers.size());
             Assert.assertEquals("column1PK", identifiers.get(0));
+        }
+    }
+
+    @Test
+    public void testRenameIdentifierWithEdgeRoles() {
+        Schema publicSchema = this.sqlgGraph.getTopology().getPublicSchema();
+        VertexLabel aVertexLabel = publicSchema.ensureVertexLabelExist("A", new HashMap<>() {{
+                    put("id1", PropertyType.varChar(10));
+                    put("id2", PropertyType.varChar(10));
+                    put("a", PropertyType.varChar(10));
+                }},
+                ListOrderedSet.listOrderedSet(List.of("id1", "id2"))
+        );
+        VertexLabel bVertexLabel = publicSchema.ensureVertexLabelExist("B", new HashMap<>() {{
+                    put("id1", PropertyType.varChar(10));
+                    put("id2", PropertyType.varChar(10));
+                    put("a", PropertyType.varChar(10));
+                }},
+                ListOrderedSet.listOrderedSet(List.of("id1", "id2"))
+        );
+        aVertexLabel.ensureEdgeLabelExist("ab", bVertexLabel);
+        this.sqlgGraph.tx().commit();
+
+        Vertex a= this.sqlgGraph.addVertex(T.label, "A", "id1", "1", "id2", "2", "a", "a");
+        Vertex b= this.sqlgGraph.addVertex(T.label, "B", "id1", "1", "id2", "2", "a", "a");
+        a.addEdge("ab", b);
+        this.sqlgGraph.tx().commit();
+
+        List<Vertex> vertices = this.sqlgGraph.traversal().V().hasLabel("A").out().toList();
+        Assert.assertEquals(1, vertices.size());
+        SchemaTable schemaTable = SchemaTable.of(this.sqlgGraph.getSqlDialect().getPublicSchema(), "A");
+        vertices = this.sqlgGraph.traversal().V().hasId(P.eq(RecordId.from(schemaTable, List.of("1", "2")))).toList();
+        Assert.assertEquals(1, vertices.size());
+
+        Optional<PropertyColumn> id1Optional = aVertexLabel.getProperty("id1");
+        Preconditions.checkState(id1Optional.isPresent());
+        ListOrderedSet<String> identifiers = aVertexLabel.getIdentifiers();
+        Assert.assertEquals(2, identifiers.size());
+        Assert.assertEquals("id1", identifiers.get(0));
+        PropertyColumn id1 = id1Optional.get();
+        Assert.assertSame(id1Optional.get(), id1);
+        id1.rename("id1PK");
+        this.sqlgGraph.tx().commit();
+
+        publicSchema = this.sqlgGraph.getTopology().getPublicSchema();
+        Optional<VertexLabel> aVertexLabelOptional = publicSchema.getVertexLabel("A");
+        Assert.assertTrue(aVertexLabelOptional.isPresent());
+        Optional<VertexLabel> bVertexLabelOptional = publicSchema.getVertexLabel("B");
+        Assert.assertTrue(bVertexLabelOptional.isPresent());
+        Optional<EdgeLabel> edgeLabelOptional = publicSchema.getEdgeLabel("ab");
+        Preconditions.checkState(edgeLabelOptional.isPresent());
+        Set<EdgeRole> outEdgeRoles = edgeLabelOptional.get().getOutEdgeRoles();
+        Assert.assertEquals(1, outEdgeRoles.size());
+        EdgeRole edgeRole = outEdgeRoles.iterator().next();
+        Assert.assertSame(aVertexLabelOptional.get(), edgeRole.getVertexLabel());
+        Set<EdgeRole> inEdgeRoles = edgeLabelOptional.get().getInEdgeRoles();
+        Assert.assertEquals(1, inEdgeRoles.size());
+        edgeRole = inEdgeRoles.iterator().next();
+        Assert.assertSame(bVertexLabelOptional.get(), edgeRole.getVertexLabel());
+
+        vertices = this.sqlgGraph.traversal().V().hasLabel("A").out().toList();
+        Assert.assertEquals(1, vertices.size());
+        schemaTable = SchemaTable.of(this.sqlgGraph.getSqlDialect().getPublicSchema(), "A");
+        vertices = this.sqlgGraph.traversal().V().hasId(P.eq(RecordId.from(schemaTable, List.of("1", "2")))).toList();
+        Assert.assertEquals(1, vertices.size());
+
+        id1Optional = aVertexLabel.getProperty("id1PK");
+        Preconditions.checkState(id1Optional.isPresent());
+        identifiers = aVertexLabel.getIdentifiers();
+        Assert.assertEquals(2, identifiers.size());
+        Assert.assertEquals("id1PK", identifiers.get(0));
+        id1 = id1Optional.get();
+        Assert.assertSame(id1Optional.get(), id1);
+        id1.rename("id1PKAgain");
+        this.sqlgGraph.tx().rollback();
+
+        if (this.sqlgGraph.getSqlDialect().supportsTransactionalSchema()) {
+            aVertexLabelOptional = this.sqlgGraph.getTopology().getPublicSchema().getVertexLabel("A");
+            Preconditions.checkState(aVertexLabelOptional.isPresent());
+            aVertexLabel = aVertexLabelOptional.get();
+            id1Optional = aVertexLabel.getProperty("id1PK");
+            Preconditions.checkState(id1Optional.isPresent());
+            identifiers = aVertexLabel.getIdentifiers();
+            Assert.assertEquals(1, identifiers.size());
+            Assert.assertEquals("id1PK", identifiers.get(0));
         }
     }