+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 ##2.1.6
 
+* Implemented [#139](https://github.com/pietermartin/sqlg/issues/139).
+Added `rename` to `VertexLabel`, `EdgeLabel` and `PropertyColumn`
 * BREAKING CHANGE: Changed `TopologyListener`'s interface, `oldValue` is now a `TopologyInf`
-*Implemented [#428](https://github.com/pietermartin/sqlg/issues/428). 
+* Implemented [#428](https://github.com/pietermartin/sqlg/issues/428). 
 Added ability to import foreign schemas and to import tables into an existing schema.
 Both import must be consistent, i.e. can not reference elements that are not being imported.